change unit tests to work with new querier handler
