+++ pkg/querier/tripperware/instantquery/instant_query_test.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
-	jsoniter "github.com/json-iterator/go"
+	"github.com/weaveworks/common/httpgrpc"
 	"io"
 	"net/http"
 	"strconv"
@@ -109,14 +109,15 @@ func TestCompressedResponse(t *testing.T) {
 	for _, tc := range []struct {
 		name        string
 		compression string
-		promBody    PrometheusInstantQueryResponse
+		jsonBody    string
+		promBody    *PrometheusInstantQueryResponse
 		status      int
 		err         error
 	}{
 		{
 			name:        "successful response",
 			compression: "gzip",
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValString.String(),
@@ -124,7 +125,7 @@ func TestCompressedResponse(t *testing.T) {
 				},
 				Headers: []*tripperware.PrometheusResponseHeader{
 					{Name: "Content-Type", Values: []string{"application/x-protobuf"}},
-					{Name: "Content-Encoding", Values: []string{"snappy"}},
+					{Name: "Content-Encoding", Values: []string{"gzip"}},
 				},
 			},
 			status: 200,
@@ -132,7 +133,7 @@ func TestCompressedResponse(t *testing.T) {
 		{
 			name:        "successful response",
 			compression: "snappy",
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValString.String(),
@@ -145,30 +146,62 @@ func TestCompressedResponse(t *testing.T) {
 			},
 			status: 200,
 		},
+		{
+			name:        `400 error`,
+			compression: `gzip`,
+			jsonBody:    `error generic 400`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, `error generic 400`),
+		},
+		{
+			name:        `400 error`,
+			compression: `snappy`,
+			jsonBody:    `error generic 400`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, `error generic 400`),
+		},
+		{
+			name:        `400 error empty body`,
+			compression: `gzip`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, ""),
+		},
+		{
+			name:        `400 error empty body`,
+			compression: `snappy`,
+			status:      400,
+			err:         httpgrpc.Errorf(400, ""),
+		},
 	} {
 		for _, c := range []bool{true, false} {
 			c := c
 			t.Run(fmt.Sprintf("%s compressed %t [%s]", tc.compression, c, tc.name), func(t *testing.T) {
 				t.Parallel()
-				protobuf, err := proto.Marshal(&tc.promBody)
-				require.NoError(t, err)
-				responseBody := bytes.NewBuffer(protobuf)
-				h := http.Header{
-					"Content-Type": []string{"application/x-protobuf"},
+				h := http.Header{}
+				var b []byte
+				if tc.promBody != nil {
+					protobuf, err := proto.Marshal(tc.promBody)
+					b = protobuf
+					require.NoError(t, err)
+					h.Set("Content-Type", "application/x-protobuf")
+				} else {
+					b = []byte(tc.jsonBody)
+					h.Set("Content-Type", "application/json")
 				}
+				responseBody := bytes.NewBuffer(b)
 
 				var buf bytes.Buffer
 				if c && tc.compression == "gzip" {
 					h.Set("Content-Encoding", "gzip")
 					w := gzip.NewWriter(&buf)
-					_, err := w.Write(protobuf)
+					_, err := w.Write(b)
 					require.NoError(t, err)
 					w.Close()
 					responseBody = &buf
 				} else if c && tc.compression == "snappy" {
 					h.Set("Content-Encoding", "snappy")
 					w := snappy.NewBufferedWriter(&buf)
-					_, err := w.Write(protobuf)
+					_, err := w.Write(b)
 					require.NoError(t, err)
 					w.Close()
 					responseBody = &buf
@@ -179,13 +212,12 @@ func TestCompressedResponse(t *testing.T) {
 					Header:     h,
 					Body:       io.NopCloser(responseBody),
 				}
-				r, err := InstantQueryCodec.DecodeResponse(context.Background(), response, nil)
+				resp, err := InstantQueryCodec.DecodeResponse(context.Background(), response, nil)
 				require.Equal(t, tc.err, err)
 
 				if err == nil {
-					resp, err := proto.Marshal(r)
 					require.NoError(t, err)
-					require.Equal(t, protobuf, resp)
+					require.Equal(t, tc.promBody, resp)
 				}
 			})
 		}
@@ -196,11 +228,11 @@ func TestResponse(t *testing.T) {
 	t.Parallel()
 	for i, tc := range []struct {
 		expectedResp string
-		promBody     PrometheusInstantQueryResponse
+		promBody     *PrometheusInstantQueryResponse
 	}{
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"string","result":[1,"foo"]}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValString.String(),
@@ -213,7 +245,7 @@ func TestResponse(t *testing.T) {
 		},
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"string","result":[1,"foo"],"stats":{"samples":{"totalQueryableSamples":10,"totalQueryableSamplesPerStep":[[1536673680,5],[1536673780,5]]}}}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValString.String(),
@@ -226,7 +258,7 @@ func TestResponse(t *testing.T) {
 		},
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"foo":"bar"},"values":[[1,"137"],[2,"137"]]}],"stats":{"samples":{"totalQueryableSamples":10,"totalQueryableSamplesPerStep":[[1536673680,5],[1536673780,5]]}}}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValMatrix.String(),
@@ -264,7 +296,7 @@ func TestResponse(t *testing.T) {
 		},
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"matrix","result":[{"metric":{"foo":"bar"},"values":[[1,"137"],[2,"137"]]}]}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValMatrix.String(),
@@ -293,7 +325,7 @@ func TestResponse(t *testing.T) {
 		},
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"scalar","result":[1,"13"]}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValString.String(),
@@ -306,7 +338,7 @@ func TestResponse(t *testing.T) {
 		},
 		{
 			expectedResp: `{"status":"success","data":{"resultType":"vector","result":[{"metric":{},"value":[1,"1266464.0146205237"]}]}}`,
-			promBody: PrometheusInstantQueryResponse{
+			promBody: &PrometheusInstantQueryResponse{
 				Status: "success",
 				Data: PrometheusInstantQueryData{
 					ResultType: model.ValVector.String(),
@@ -332,13 +364,7 @@ func TestResponse(t *testing.T) {
 		tc := tc
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			t.Parallel()
-
-			json = jsoniter.Config{
-				EscapeHTML:             false, // No HTML in our responses.
-				SortMapKeys:            true,
-				ValidateJsonRawMessage: false,
-			}.Froze()
-			protobuf, err := proto.Marshal(&tc.promBody)
+			protobuf, err := proto.Marshal(tc.promBody)
 			require.NoError(t, err)
 
 			response := &http.Response{
@@ -993,7 +1019,7 @@ func TestMergeResponse(t *testing.T) {
 									SampleStreams: []tripperware.SampleStream{
 										{
 											Labels: []cortexpb.LabelAdapter{
-												{"__name__", "foo"},
+												{"__name__", "bar"},
 											},
 											Samples: []cortexpb.Sample{
 												{Value: 1, TimestampMs: 1000},
@@ -1019,7 +1045,7 @@ func TestMergeResponse(t *testing.T) {
 									SampleStreams: []tripperware.SampleStream{
 										{
 											Labels: []cortexpb.LabelAdapter{
-												{"__name__", "bar"},
+												{"__name__", "foo"},
 											},
 											Samples: []cortexpb.Sample{
 												{Value: 3, TimestampMs: 3000},
